requirements.txt
# 游戏字幕朗读工具 - 依赖包列表

# ============ 核心依赖 ============
# OpenAI SDK（用于 Qwen 全模态 API）
openai>=1.0.0

# 音频处理
pyaudio>=0.2.11
numpy>=1.24.0
soundfile>=0.12.0

# 屏幕截图（高性能）
mss>=9.0.0

# 图像处理
Pillow>=10.0.0

# 全局快捷键（无需管理员权限）
pynput>=1.7.6

# 环境变量管理
python-dotenv>=1.0.0
